+++ protected/humhub/modules/user/authclient/AuthClientHelpers.php
@@ -36,10 +36,7 @@ class AuthClientHelpers
             /**
              * @var interfaces\PrimaryClient $authClient
              */
-            return User::findOne([
-                        $authClient->getUserTableIdAttribute() => $attributes['id'],
-                        'auth_mode' => $authClient->getId()
-            ]);
+            return $authClient->getUser();
         }
 
         $auth = Auth::find()->where(['source' => $authClient->getId(), 'source_id' => $attributes['id']])->one();
@@ -122,7 +119,7 @@ class AuthClientHelpers
             $attributes = $authClient->getUserAttributes();
             foreach ($authClient->getSyncAttributes() as $attributeName) {
                 if (isset($attributes[$attributeName])) {
-                    if (in_array($attributeName, ['email', 'username'])) {
+                    if (in_array($attributeName, ['email', 'username', 'authclient_id'])) {
                         $user->setAttribute($attributeName, $attributes[$attributeName]);
                     } else {
                         $user->profile->setAttribute($attributeName, $attributes[$attributeName]);
@@ -165,7 +162,7 @@ class AuthClientHelpers
         $registration = new \humhub\modules\user\models\forms\Registration();
         $registration->enablePasswordForm = false;
         $registration->enableEmailField = true;
-        
+
         if ($authClient instanceof interfaces\ApprovalBypass) {
             $registration->enableUserApproval = false;
         }